moved singleton onboarding
