+++ app/src/androidTest/java/com/sidgowda/pawcalc/navigation/NavControllerTest.kt
@@ -1,0 +1,4 @@
+package com.sidgowda.pawcalc.navigation
+
+class NavControllerTest {
+}
